Preserve the original enexSources option value
Move code dealing with options.enexSources down a level into same file
where it is iterated over meaning we can store the collection to loop
over in a variable rather than modifying the options original options
value, thereby overwriting (losing) the user-configured collection.

This means that the log reports the folders configured by the user, not
the final set of files that the program iterates over.